+++ tracer/src/Datadog.Tracer.Native/Datadog.Tracer.Native.vcxproj
@@ -269,6 +269,11 @@
     <ClInclude Include="..\..\..\shared\src\native-src\miniutfdata.h" />
     <ClInclude Include="..\..\..\shared\src\native-src\pal.h" />
     <ClInclude Include="..\..\..\shared\src\native-src\string.h" />
+    <ClInclude Include="fault_tolerant_tokens.h" />
+    <ClInclude Include="fault_tolerant_envionrment_variables_util.h" />
+    <ClInclude Include="fault_tolerant_environment_variables.h" />
+    <ClInclude Include="fault_tolerant_manager.h" />
+    <ClInclude Include="fault_tolerant_rewriter.h" />
     <ClInclude Include="fault_tolerant_tracker.h" />
     <ClInclude Include="iast\app_domain_info.h" />
     <ClInclude Include="iast\aspect.h" />
@@ -337,6 +342,10 @@
     <ClCompile Include="debugger_rejit_handler_module_method.cpp" />
     <ClCompile Include="debugger_rejit_preprocessor.cpp" />
     <ClCompile Include="environment_variables_util.cpp" />
+    <ClCompile Include="fault_tolerant_tokens.cpp" />
+    <ClCompile Include="fault_tolerant_envionrment_variables_util.cpp" />
+    <ClCompile Include="fault_tolerant_manager.cpp" />
+    <ClCompile Include="fault_tolerant_rewriter.cpp" />
     <ClCompile Include="fault_tolerant_tracker.cpp" />
     <ClCompile Include="iast\aspect.cpp" />
     <ClCompile Include="iast\aspect_filter.cpp" />
