+++ tracer/src/Datadog.Tracer.Native/debugger_constants.h
@@ -64,6 +64,8 @@ const WSTRING invalid_probe_type_is_by_ref_like =
     WStr("Dynamic Instrumentation of methods in a `ref-struct` is not yet supported.");
 const WSTRING non_supported_compiled_bytecode =
     WStr("Compiled code with `tail` call is not yet supported (F#).");
+const WSTRING type_contains_invalid_symbol = 
+    WStr("The type is not supported.");
 const WSTRING invalid_probe_failed_to_instrument_method_probe = 
     GetGenericErrorMessageWithErrorCode(1);
 const WSTRING invalid_probe_failed_to_instrument_line_probe =
