+++ tracer/src/Datadog.Tracer.Native/debugger_environment_variables_util.cpp
@@ -14,9 +14,4 @@ bool IsDebuggerInstrumentAllEnabled()
     CheckIfTrue(GetEnvironmentValue(environment::internal_instrument_all_enabled));
 }
 
-bool IsFaultTolerantInstrumentationEnabled()
-{
-    CheckIfTrue(GetEnvironmentValue(environment::fault_tolerant_instrumentation_enabled));
-}
-
 } // namespace debugger
