+++ tracer/src/Datadog.Tracer.Native/debugger_environment_variables_util.h
@@ -10,7 +10,6 @@ namespace debugger
 
 bool IsDebuggerEnabled();
 bool IsDebuggerInstrumentAllEnabled();
-bool IsFaultTolerantInstrumentationEnabled();
 
 } // namespace debugger
