+++ tracer/src/Datadog.Tracer.Native/debugger_probes_instrumentation_requester.h
@@ -5,6 +5,14 @@
 #include "../../../shared/src/native-src/string.h"
 #include <corprof.h>
 #include "debugger_members.h"
+#include "fault_tolerant_manager.h"
+
+// forward declaration
+
+namespace fault_tolerant
+{
+class FaultTolerantManager;
+} // namespace fault_tolerant
 
 namespace debugger
 {
@@ -18,9 +26,8 @@ private:
     std::unique_ptr<DebuggerRejitPreprocessor> m_debugger_rejit_preprocessor = nullptr;
     std::shared_ptr<RejitHandler> m_rejit_handler = nullptr;
     std::shared_ptr<RejitWorkOffloader> m_work_offloader = nullptr;
+    std::shared_ptr<fault_tolerant::FaultTolerantManager> m_fault_tolerant_manager = nullptr;
     bool is_debugger_enabled = false;
-    bool is_fault_tolerant_instrumentation_enabled = false;
-    bool flag = true;
 
     static bool IsCoreLibOr3rdParty(const WSTRING& assemblyName);
     static std::wstring GenerateRandomProbeId();
@@ -32,17 +39,14 @@ private:
                          std::set<trace::MethodIdentifier>& rejitRequests);
     void AddLineProbes(debugger::DebuggerLineProbeDefinition* lineProbes, int lineProbesLength,
                        std::set<MethodIdentifier>& rejitRequests);
-    void DetermineReInstrumentProbes(std::set<MethodIdentifier>& revertRequests, std::set<MethodIdentifier>& reInstrumentRequests) const;
+    void DetermineReInstrumentProbes(std::set<MethodIdentifier>& revertRequests,
+                                     std::set<MethodIdentifier>& reInstrumentRequests) const;
 
     bool ProbeIdExists(const WCHAR* probeId);
 
-    void HandleFaultTolerantInstrumentationIfEnabled(ModuleID moduleId, const ModuleInfo& moduleInfo,
-                                            ComPtr<IMetaDataImport2> metadataImport,
-                                                     ComPtr<IMetaDataEmit2> metadataEmit, mdTypeDef typeDef) const;
-
 public:
     DebuggerProbesInstrumentationRequester(CorProfiler* corProfiler, std::shared_ptr<trace::RejitHandler> rejit_handler,
-                                           std::shared_ptr<trace::RejitWorkOffloader> work_offloader);
+                                           std::shared_ptr<trace::RejitWorkOffloader> work_offloader, std::shared_ptr<fault_tolerant::FaultTolerantManager> fault_tolerant_manager);
 
     void InstrumentProbes(debugger::DebuggerMethodProbeDefinition* methodProbes, int methodProbesLength,
                    debugger::DebuggerLineProbeDefinition* lineProbes, int lineProbesLength,
