+++ tracer/src/Datadog.Tracer.Native/debugger_rejit_preprocessor.cpp
@@ -252,14 +252,16 @@ DebuggerRejitPreprocessor::CreateMethod(const mdMethodDef methodDef, RejitHandle
                                         const FunctionInfo& functionInfo,
                                         const std::shared_ptr<MethodProbeDefinition>& methodProbe)
 {
-    return std::make_unique<DebuggerRejitHandlerModuleMethod>(methodDef, module, functionInfo, std::make_unique<DebuggerMethodRewriter>(m_corProfiler));
+    auto faultTolerantRewriter = std::make_unique<fault_tolerant::FaultTolerantRewriter>(m_corProfiler, std::make_unique<DebuggerMethodRewriter>(m_corProfiler));
+    return std::make_unique<DebuggerRejitHandlerModuleMethod>(methodDef, module, functionInfo, std::move(faultTolerantRewriter));
 }
 
 const std::unique_ptr<RejitHandlerModuleMethod>
 DebuggerRejitPreprocessor::CreateMethod(const mdMethodDef methodDef, RejitHandlerModule* module,
                                         const FunctionInfo& functionInfo) const
 {
-    return std::make_unique<DebuggerRejitHandlerModuleMethod>(methodDef, module, functionInfo, std::make_unique<DebuggerMethodRewriter>(m_corProfiler));
+    auto faultTolerantRewriter = std::make_unique<fault_tolerant::FaultTolerantRewriter>(m_corProfiler, std::make_unique<DebuggerMethodRewriter>(m_corProfiler));
+    return std::make_unique<DebuggerRejitHandlerModuleMethod>(methodDef, module, functionInfo, std::move(faultTolerantRewriter));
 }
 
 void DebuggerRejitPreprocessor::UpdateMethod(RejitHandlerModuleMethod* methodHandler, const std::shared_ptr<MethodProbeDefinition>& methodProbe)
