+++ tracer/src/Datadog.Tracer.Native/fault_tolerant_envionrment_variables_util.h
@@ -1,0 +1,11 @@
+#ifndef DD_CLR_PROFILER_FAULT_TOLERANT_ENVIRONMENT_VARIABLES_UTIL_H_
+#define DD_CLR_PROFILER_FAULT_TOLERANT_ENVIRONMENT_VARIABLES_UTIL_H_
+
+namespace fault_tolerant
+{
+
+bool IsFaultTolerantInstrumentationEnabled();
+
+} // namespace fault_tolerant
+
+#endif
