+++ tracer/src/Datadog.Tracer.Native/fault_tolerant_environment_variables.cpp
@@ -1,0 +1,1 @@
+#include "fault_tolerant_environment_variables.h"
