+++ tracer/test/test-applications/debugger/Samples.Probes/Program.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
 using Samples.Probes;
@@ -61,7 +60,7 @@ public static class Program
                 {
                     run.Run();
                 }
-                catch (Exception ex) when (ex.Source == "ASD")
+                catch
                 {
                     // Ignored
                 }
